new updates

+++ src/Main.java
@@ -1,8 +1,8 @@
 public class Main {
     public static void main(String[] args) {
-        int balance = 2_000_000_000;
-        int transfer = 500_000_000;
-        int total = balance + transfer;
+        long balance = 2_000_000_000;
+        long transfer = 500_000_000;
+        long total = balance + transfer;
       System.out.println(total);
     }
 }
